refactor e2e test udp
