+++ ownserver_test/src/lib.rs
@@ -1,0 +1,305 @@
+use std::sync::Arc;
+use once_cell::sync::OnceCell;
+
+use ownserver_auth::build_routes;
+use ownserver_server::{
+    proxy_server,
+    Config,
+    Store,
+};
+use ownserver_lib::{EndpointClaim, Protocol};
+use ownserver::{
+    proxy_client::{self, ClientInfo},
+    Store as ClientStore,
+};
+use tokio::io::{AsyncReadExt, AsyncWriteExt};
+use tokio::net::TcpListener;
+use tokio_util::sync::CancellationToken;
+use tokio::net::UdpSocket;
+use futures::Future;
+
+
+#[macro_export]
+macro_rules! wait {
+    () => {
+        tokio::time::sleep(std::time::Duration::from_secs(3)).await;
+        
+    };
+}
+
+
+#[macro_export]
+macro_rules! assert_tcp_socket_bytes_matches {
+    ($socket:expr, $expected:expr) => {
+        let mut buf = [0; 4 * 1024];
+        let n = $socket
+            .read(&mut buf)
+            .await
+            .expect("failed to read data from socket");
+        let data = buf[..n].to_vec();
+
+        assert_eq!(data, $expected);  
+    };
+}
+
+#[macro_export]
+macro_rules! assert_udp_socket_bytes_matches {
+    ($socket:expr, $expected:expr) => {
+        let mut buf = [0; 4 * 1024];
+        let n = $socket
+            .recv(&mut buf)
+            .await
+            .expect("failed to read data from socket");
+        let data = buf[..n].to_vec();
+
+        assert_eq!(data, $expected);  
+    };
+}
+
+pub static CONFIG: OnceCell<Config> = OnceCell::new();
+
+pub const CONTROL_PORT: u16 = 5000;
+pub const LOCAL_PORT: u16 = 3000;
+pub const REMOTE_PORT_START: u16 = 4500;
+pub const REMOTE_PORT_END: u16 = 4599;
+pub const TOKEN_PORT: u16 = 8888;
+
+
+pub struct TokenServer {
+}
+
+
+pub struct ProxyServer {
+    pub store: Arc<Store>,
+}
+
+pub struct ProxyClient {
+    pub client_info: ClientInfo,
+    pub cancellation_token: CancellationToken,
+}
+
+    
+pub struct LocalServer {
+    
+}
+
+pub async fn launch_token_server(token_port: u16) -> TokenServer {
+    let hosts = vec![
+        "127.0.0.1".to_string(),
+    ];
+    let routes = build_routes("supersecret".to_string(), hosts);
+
+    tokio::spawn(async move {
+        warp::serve(routes).run(([127, 0, 0, 1], token_port)).await;
+    });
+
+    wait!();
+    TokenServer {}
+}
+
+
+pub async fn launch_proxy_server(
+    control_port: u16,
+    remote_port_start: u16,
+    remote_port_end: u16
+) -> Result<ProxyServer, Box<dyn std::error::Error>> {
+
+    let config = CONFIG.get_or_init(||
+        Config {
+            control_port,
+            token_secret: "supersecret".to_string(),
+            host: "127.0.0.1".to_string(),
+            remote_port_start,
+            remote_port_end,
+            periodic_cleanup_interval: 2 << 30,
+            periodic_ping_interval: 2 << 30,
+        }
+    );
+
+    let store = Arc::new(Store::new(config.remote_port_start..config.remote_port_end));
+
+    let store_ = store.clone();
+    tokio::spawn(async move {
+        proxy_server::run(
+            &CONFIG,
+            store_,
+        )
+        .await.join_next().await;
+    });
+
+    wait!();
+    Ok(ProxyServer {
+        store,
+    })
+}
+
+
+pub mod tcp {
+    use super::*;
+    
+    pub async fn launch_proxy_client(
+        control_port: u16,
+        local_port: u16,
+    ) -> Result<ProxyClient, Box<dyn std::error::Error>> {
+        let client_store: Arc<ClientStore> = Default::default();
+        let cancellation_token = CancellationToken::new();
+    
+        let endpoint_claims = vec![EndpointClaim {
+            protocol: Protocol::TCP,
+            local_port,
+            remote_port: 0,
+        }];
+    
+        let (client_info, mut set) =
+                proxy_client::run(client_store, control_port, "http://127.0.0.1:8888/v0/request_token", cancellation_token.clone(), endpoint_claims)
+                    .await
+                    .expect("failed to launch proxy_client");
+        tokio::spawn(async move {
+            while let Some(res) = set.join_next().await {
+                let _ = res.unwrap();
+            }
+        });
+    
+        wait!();
+        Ok(ProxyClient {
+            client_info,
+            cancellation_token,
+        })
+    }
+
+    pub async fn launch_local_server(local_port: u16) -> LocalServer {
+        let local_server = async move {
+            let listener = TcpListener::bind(format!("127.0.0.1:{}", local_port))
+                .await
+                .unwrap();
+    
+            loop {
+                let (mut socket, _) = listener.accept().await.expect("No connections to accept");
+    
+                tokio::spawn(async move {
+                    loop {
+                        let mut buf = [0; 4 * 1024];
+                        let n = socket
+                            .read(&mut buf)
+                            .await
+                            .expect("failed to read data from socket");
+                        if n == 0 {
+                            return;
+                        }
+    
+                        let mut msg = b"hello, ".to_vec();
+                        msg.append(&mut buf[..n].to_vec());
+                        socket
+                            .write_all(&msg)
+                            .await
+                            .expect("failed to write packet data to local tcp socket");
+                    }
+                });
+            }
+        };
+        tokio::spawn(local_server);
+    
+        wait!();
+        LocalServer {}
+    }
+
+
+    pub async fn launch_all<T>(test_func: impl FnOnce(TokenServer, ProxyServer, LocalServer, ProxyClient) -> T)
+        where
+        T: Future<Output = Result<(), Box<dyn std::error::Error>>> + Send,
+    {
+        let token_server = launch_token_server(TOKEN_PORT).await;
+        wait!();
+
+        let proxy_server = launch_proxy_server(CONTROL_PORT, REMOTE_PORT_START, REMOTE_PORT_END).await.expect("failed to launch proxy server");
+        let local_server = launch_local_server(LOCAL_PORT).await;
+
+        wait!();
+        let proxy_client = launch_proxy_client(CONTROL_PORT, LOCAL_PORT).await.expect("failed to launch proxy client");
+
+        test_func(token_server, proxy_server, local_server, proxy_client).await.expect("failed to call test_func");
+    }
+}
+
+pub mod udp {
+    use super::*;
+
+    pub async fn launch_proxy_client(
+        control_port: u16,
+        local_port: u16,
+    ) -> Result<ProxyClient, Box<dyn std::error::Error>> {
+        let client_store: Arc<ClientStore> = Default::default();
+        let cancellation_token = CancellationToken::new();
+    
+        let endpoint_claims = vec![EndpointClaim {
+            protocol: Protocol::UDP,
+            local_port,
+            remote_port: 0,
+        }];
+        let (client_info, mut set) =
+            proxy_client::run(client_store, control_port, "http://127.0.0.1:8888/v0/request_token", cancellation_token.clone(), endpoint_claims)
+                .await
+                .expect("failed to launch proxy_client");
+        tokio::spawn(async move {
+            while let Some(res) = set.join_next().await {
+                let _ = res.unwrap();
+            }
+        });
+    
+        wait!();
+        Ok(ProxyClient {
+            client_info,
+            cancellation_token,
+        })
+    }
+    
+    pub async fn launch_local_server(local_port: u16) -> LocalServer {
+        let local_server = async move {
+            let socket = UdpSocket::bind(format!("127.0.0.1:{}", local_port)).await.unwrap();
+    
+            tokio::spawn(async move {
+                loop {
+                    let mut buf = [0; 4 * 1024];
+                    let (n, addr) = socket
+                        .recv_from(&mut buf)
+                        .await
+                        .expect("failed to read data from socket");
+                    if n == 0 {
+                        return;
+                    }
+    
+    
+                    let mut msg = b"hello, ".to_vec();
+                    msg.append(&mut buf[..n].to_vec());
+                    tracing::info!("send data to remote socket: {:?}", msg);
+                    socket
+                        .send_to(&msg, addr)
+                        .await
+                        .expect("failed to write packet data to local udp socket");
+                }
+            });
+        };
+        tokio::spawn(local_server);
+
+        wait!();
+        LocalServer {}
+    }
+
+
+    pub async fn launch_all<T>(test_func: impl FnOnce(TokenServer, ProxyServer, LocalServer, ProxyClient) -> T)
+        where
+        T: Future<Output = Result<(), Box<dyn std::error::Error>>> + Send,
+    {
+        let token_server = launch_token_server(TOKEN_PORT).await;
+        wait!();
+
+        let proxy_server = launch_proxy_server(CONTROL_PORT, REMOTE_PORT_START, REMOTE_PORT_END).await.expect("failed to launch proxy server");
+        let local_server = launch_local_server(LOCAL_PORT).await;
+
+        wait!();
+        let proxy_client = launch_proxy_client(CONTROL_PORT, LOCAL_PORT).await.expect("failed to launch proxy client");
+
+        test_func(token_server, proxy_server, local_server, proxy_client).await.expect("failed to call test_func");
+    }
+}
+
