+++ ownserver_test/tests/lib.rs
@@ -1,1 +1,1 @@
-pub mod server;
+pub mod e2e_test;
